@@ -1,8 +1,8 @@
 source 'https://rubygems.org'
-ruby '2.3.3'
+ruby '2.7.2'
 gem 'dotenv-rails', groups: [:development]
-gem 'rails', '~> 5.0.3'
-gem 'puma', '~> 3.12'
+gem 'rails', '~> 5.2.6'
+gem 'puma'
 gem 'sass-rails', '~> 5.0'
 gem 'uglifier', '>= 1.3.0'
 gem 'coffee-rails', '~> 4.2'
