@@ -1,6 +1,6 @@
 class JuniferController < ApplicationController
-  skip_before_filter :verify_authenticity_token
-  before_filter :restrict_access
+  skip_before_action :verify_authenticity_token
+  before_action :restrict_access
 
   def proxy
     baseurl = "#{ENV['JUNIFER_PROTOCOL']}://#{ENV['JUNIFER_HOSTNAME']}:#{ENV['JUNIFER_PORT']}"
@@ -24,7 +24,7 @@ class JuniferController < ApplicationController
 
   def restrict_access
     authenticate_or_request_with_http_token do |token, options|
-      token == ENV.fetch('API_AUTH_KEY_JUNIFER')
+      token == ENV.fetch('API_AUTH_KEY_JUNIFER') || token == ENV.fetch('NEW_API_AUTH_KEY_JUNIFER')
     end
   end
 end
